+++ android/src/main/kotlin/com/solusibejo/screen_time/const/MethodName.kt
@@ -5,7 +5,6 @@ object MethodName {
     const val installedApps = "installedApps"
     const val appUsageData = "appUsageData"
     const val monitoringAppUsage = "monitoringAppUsage"
-    const val openAccessibilitySettings = "openAccessibilitySettings"
-    const val isAppMonitoringServiceEnabled = "isAppMonitoringServiceEnabled"
     const val configureAppMonitoringService = "configureAppMonitoringService"
+    const val permissionStatus = "permissionStatus"
 }