+++ lib/src/const/argument.dart
@@ -11,4 +11,5 @@ class Argument {
 
   static const String lookbackTimeMs = 'lookbackTimeMs';
   static const String interval = 'interval';
+  static const String permissionType = 'permissionType';
 }