+++ lib/src/model/screen_time_permission_type.dart
@@ -1,0 +1,1 @@
+enum ScreenTimePermissionType { appUsage, accessibilitySettings }